requirements.txt
asgiref==3.4.1
asyncpg==0.23.0
backcall==0.2.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
databases==0.4.3
decorator==5.0.9
ecdsa==0.17.0
fastapi==0.66.0
h11==0.12.0
httptools==0.2.0
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
python-dotenv==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
